refactor: split ptoas.cpp into four translation units
Resolve the deferred huge_non_headerfile finding (#89): ptoas.cpp
(3527 nbnc) is split along its existing section boundaries, moving code
verbatim into three new TUs coordinated by an internal header.

- ptoas.cpp (~333 nbnc): inliner/emitc compat passes, public registration
  APIs, and the command-line option definitions
- ptoas_name_hints.cpp (~826): textual SSA name recovery, provenance
  annotation, hint-marker decoding/strip
- ptoas_cpp_rewrite.cpp (~1208): call-site marker rewriting, EmitC integer
  attribute normalization, scalar-GM flush, verbatim repair, constant-decl
  hoisting
- ptoas_pipeline.cpp (~1399): arch resolution, build-level validation,
  auto-sync/frontend passes, backend pipelines, compilePTOASModule
- ptoas_internal.h: shared inline capacities, the StringRefVector and
  FunctionBlockArgHintMap aliases, extern declarations for the cl::opt
  globals consumed by the pipeline TU, and the cross-TU entry points

Cross-TU symbols either moved with their only caller (arch helpers, build
validation, seam-IR printers, reorderEmitCFunctions, pass structs) or lost
'static' for a header declaration (name-hint and rewrite entry points).
Thirteen unbraced control bodies in the moved code were braced; the
remaining G.FMT.11 prefilter hits are false positives on already-braced
multi-paren conditions, and the one assert is a debug-only internal
invariant per the rule-resolution policy.

Validated on 144 (LLVM 19.1.7): full rebuild of PTOASCompiler,
PTOASPythonCore, PTODSLPackage, PTOASPythonPackage, ptobc; ctest 45/45
(PTODSL+PTO), 23/23 ptobc, lit check-pto 1843 passed / 1 unsupported;
smoke covers --emit-pto-ir, EmitC C++, and VPTO emission.
